Container_App/requirements.txt
adal==1.2.2
azure-common==1.1.23
azure-servicebus==0.50.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
entrypoints==0.3
flake8==3.7.9
idna==2.8
imageio==2.6.1
isodate==0.6.0
kiwisolver==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
msrest==0.6.10
msrestazure==0.6.2
networkx==2.4
numpy==1.17.4
oauthlib==3.1.0
Pillow==6.2.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.5
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.22.0
requests-oauthlib==1.3.0
scikit-image==0.16.2
scipy==1.4.0
six==1.13.0
uamqp==1.2.5
urllib3==1.25.7
